refactor(totalText): refactor TotalText component
