+++ src/components/Statistics/YearPanel/YearPanel.tsx
@@ -7,9 +7,8 @@ import { useAppSelector } from '../../../app/hooks';
 import { TRACKERS_BEGIN_IN } from '../../../config/Constants';
 import Tracker from '../../../models/Tracker';
 import { selectYearEntries } from '../../../store/trackers/trackers.selectors';
-import { getAggregatedCompletions } from '../../../store/trackers/utils';
 import YearChart from '../../charts/YearChart/YearChart';
-import TotalText from './TotalText';
+import TotalText from '../TotalText';
 
 interface Props {
   beginDate: Date;
@@ -18,7 +17,7 @@ interface Props {
 }
 
 const YearPanel: FC<Props> = ({ beginDate, setBeginDate, tracker }) => {
-  const monthEntries = useAppSelector((state) => selectYearEntries(state, beginDate, tracker.id));
+  const yearEntries = useAppSelector((state) => selectYearEntries(state, beginDate, tracker.id));
 
   const handleOnYearChange = (date: unknown) => {
     if (date) setBeginDate(startOfYear(date as Date));
@@ -38,9 +37,12 @@ const YearPanel: FC<Props> = ({ beginDate, setBeginDate, tracker }) => {
         />
       </Box>
       <Box sx={{ mb: 1 }}>
-        <TotalText completions={getAggregatedCompletions(monthEntries)} />
+        <TotalText
+          entries={yearEntries}
+          noCompletionsText="Il n'y a pas eu d'activité durant le mois pour le tracker sélectionné."
+        />
       </Box>
-      {monthEntries.length > 0 && <YearChart beginDate={beginDate} entries={monthEntries} />}
+      {yearEntries.length > 0 && <YearChart beginDate={beginDate} entries={yearEntries} />}
     </>
   );
 };
